backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.6
requests>=2.28.2
python-dotenv>=1.0.0
pydantic>=1.10.7
pydantic-settings==2.1.0
aiohttp>=3.8.4
pytest>=7.3.1 